Now call me lord of android

+++ app/src/main/java/com/example/homeworkskotlin/response/WeatherResponse.kt
@@ -89,5 +89,5 @@ data class Wind(
     @SerializedName("gust")
     var gust: Int,
     @SerializedName("speed")
-    var speed: Int
+    var speed: Double
 )
